# reservas-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservas API",
        "description": "Documentación de la API de reservas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Listar reservas",
                "description": "Devuelve las reservas filtradas por usuario y rango de fechas.",
                "operationId": "95164680d4eadc11331b0f3e74f590a9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha inicial del rango",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-12"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha final del rango",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "service_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "fecha_inicio": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-06-15T10:00:00"
                                            },
                                            "fecha_fin": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-06-15T11:00:00"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Crear reserva",
                "description": "Permite a un usuario crear una reserva para un servicio, validando horarios y reglas de anticipación.",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "service_id",
                                    "fecha_inicio"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-15T10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "fecha_inicio": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-06-15T10:00:00"
                                        },
                                        "fecha_fin": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-06-15T11:00:00"
                                        },
                                        "estado": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (ej. menos de 2 horas de anticipación)"
                    }
                }
            }
        },
        "/api/reservations/{id}/cancel": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Cancelar reserva",
                "description": "Cancela una reserva activa y devuelve el monto de reembolso según las reglas del servicio y plan del usuario.",
                "operationId": "2f16eafaf87daf9a39f3ede6185468a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva a cancelar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada con reembolso calculado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "reembolso": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Listar servicios",
                "description": "Devuelve todos los servicios disponibles con sus datos básicos.",
                "operationId": "4c175c5d6594b4a5112bbe15a46e2bb4",
                "responses": {
                    "200": {
                        "description": "Lista de servicios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Corte de cabello"
                                            },
                                            "duration": {
                                                "type": "integer",
                                                "example": 60
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50000
                                            },
                                            "non_refundable": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear servicio",
                "description": "Crea un nuevo servicio con nombre, duración, precio y si es reembolsable o no.",
                "operationId": "b05b71c30267faa468172458580fb6fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "duration",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Corte de cabello"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "non_refundable": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Corte de cabello"
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 60
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50000
                                        },
                                        "non_refundable": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Devuelve todos los usuarios con sus datos básicos y reservas activas.",
                "operationId": "1ad1cf4ffbfc84b02284335bdd828008",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "juan@example.com"
                                            },
                                            "plan": {
                                                "type": "string",
                                                "example": "standard"
                                            },
                                            "reservas_activas": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Reservas",
            "description": "Reservas"
        },
        {
            "name": "Servicios",
            "description": "Servicios"
        },
        {
            "name": "Usuarios",
            "description": "Usuarios"
        }
    ]
}